Castor3D: Split of LightingPass to multiple files.
